fix: silence MSVC C4267 narrowing on size_t→Index containers (3 sites)
Windows LLP64 makes size_t = 64-bit and Index/unsigned int = 32-bit, so
implicit narrowing from container.size() into 32-bit container element
types triggers C4267. The values in question never approach INT32_MAX
in any real workload — these are per-block / per-snapshot counts — so
explicit static_cast is the right minimal fix.

Sites (all surfaced by CI run 25968675219 Windows log):
- CIFFile.C:447 — std::pair<String, Index>(category, data.size()-1)
- binaryFingerprintMethods.C:1618 — multimap<uint, uint>::insert from
  make_pair(ccs_iter->second.size(), ...)
- conformationSet.C:46 — Conformation(snapshot_order_.size(), score)
  where Conformation = pair<Index, float>

DCDFile.C:1 C4828 (illegal source char at offset 0xa51) reported by
the same run is currently NOT reproducible — file is clean ASCII on
HEAD (LF endings, no BOM, no high bytes). May have been a stale
checkout or autogen-include artifact. Re-checking on next CI run.

Local verification: macOS BALL build clean.

+++ source/STRUCTURE/binaryFingerprintMethods.C
@@ -1615,7 +1615,12 @@ bool BinaryFingerprintMethods::connectedComponents(const vector<unsigned int>& s
 		multimap<unsigned int, unsigned int>::iterator size_iter;
 		for (ccs_iter=ccs_tmp.begin(); ccs_iter!=ccs_tmp.end(); ++ccs_iter)
 		{
-			cc_sizes.insert(make_pair(ccs_iter->second.size(), ccs_iter->first));
+			// Explicit narrowing cast from size_t → unsigned int to satisfy
+			// the multimap<unsigned int, unsigned int> key type and silence
+			// MSVC C4267 (Windows LLP64: size_t is 64-bit, unsigned int is
+			// 32-bit). CC sizes never exceed UINT_MAX in practice — the
+			// container fits in process memory long before this overflows.
+			cc_sizes.insert(make_pair(static_cast<unsigned int>(ccs_iter->second.size()), ccs_iter->first));
 		}
 		
 		// STEP 3: Write information in return data structures
